remove unused files; minor modifications to car datatype; move car data maanger to cardata directory
