+++ warehouse/query-core/src/test/resources/input/ford-cars.csv
@@ -1,4 +1,3 @@
-# Date, EventID, Make, Model, Color, Doors, Wheels, Desc
-20170501,ford-eventid-001,ford,explorer,blue,4,4,common description
-20170601,ford-eventid-002,ford,mustang,silver,2,4,common description
-20170701,ford-eventid-003,ford,fusion,magenta,4,4,common description
+20150501,ford-eventid-001,ford,explorer,blue,4,4,common description
+20150601,ford-eventid-002,ford,mustang,silver,2,4,common description
+20150701,ford-eventid-003,ford,fusion,magenta,4,4,common description
