+++ warehouse/query-core/src/test/resources/input/tesla-cars.csv
@@ -1,4 +1,3 @@
-# Date, EventID, Make, Model, Color, Doors, Wheels, Desc
-20170501,tesla-eventid-001,tesla,models,red,4,4,common description
-20170601,tesla-eventid-002,tesla,modelx,blue,4,4,common description
-20170701,tesla-eventid-003,tesla,model3,yellow,4,4,common description
+20150501,tesla-eventid-001,tesla,models,red,4,4,common description
+20150601,tesla-eventid-002,tesla,modelx,blue,4,4,common description
+20150701,tesla-eventid-003,tesla,model3,yellow,4,4,common description
